slackco-backend/requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-cors-headers==1.1.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
psycopg2==2.6.1
PyJWT==1.4.0
static3==0.7.0
